Content Model Shadow Edit: Refactor Editor class
