+++ packages/roosterjs-editor-core/lib/editor/Editor.ts
@@ -1,16 +1,17 @@
+import { createEditorCore } from './createEditorCore';
 import { EditorBase } from './EditorBase';
-import { EditorOptions } from 'roosterjs-editor-types';
+import { EditorCore, EditorOptions } from 'roosterjs-editor-types';
 
 /**
  * RoosterJs core editor class
  */
-export default class Editor extends EditorBase {
+export default class Editor extends EditorBase<EditorCore> {
     /**
-     * Creates an instance of Editor
+     * Creates an instance of EditorBase
      * @param contentDiv The DIV HTML element which will be the container element of editor
      * @param options An optional options object to customize the editor
      */
     constructor(contentDiv: HTMLDivElement, options: EditorOptions = {}) {
-        super(contentDiv, options);
+        super(contentDiv, options, createEditorCore);
     }
 }
