+++ packages/roosterjs-editor-core/lib/index.ts
@@ -1,3 +1,5 @@
 // Classes
 export { default as Editor } from './editor/Editor';
 export { EditorBase } from './editor/EditorBase';
+export { isFeatureEnabled } from './editor/isFeatureEnabled';
+export { createEditorCore } from './editor/createEditorCore';
