+++ packages/roosterjs-editor-types/lib/interface/EditorCore.ts
@@ -4,6 +4,7 @@ import DarkColorHandler from './DarkColorHandler';
 import EditorPlugin from './EditorPlugin';
 import NodePosition from './NodePosition';
 import Rect from './Rect';
+import SelectionPath from './SelectionPath';
 import TableSelection from './TableSelection';
 import { ChangeSource } from '../enum/ChangeSource';
 import { ColorTransformDirection } from '../enum/ColorTransformDirection';
@@ -15,6 +16,7 @@ import { InsertOption } from './InsertOption';
 import { PendableFormatState, StyleBasedFormatState } from './FormatState';
 import { PluginEvent } from '../event/PluginEvent';
 import { PluginState } from './CorePlugins';
+import { PositionType } from '../enum/PositionType';
 import { SizeTransformer } from '../type/SizeTransformer';
 import { TableSelectionRange } from './SelectionRangeEx';
 import { TrustedHTMLHandler } from '../type/TrustedHTMLHandler';
@@ -215,6 +217,23 @@ export type InsertNode = (core: EditorCore, node: Node, option: InsertOption | n
 export type RestoreUndoSnapshot = (core: EditorCore, step: number) => void;
 
 /**
+ * Select content according to the given information.
+ * There are a bunch of allowed combination of parameters. See IEditor.select for more details
+ * @param core The editor core object
+ * @param arg1 A DOM Range, or SelectionRangeEx, or NodePosition, or Node, or Selection Path
+ * @param arg2 (optional) A NodePosition, or an offset number, or a PositionType, or a TableSelection
+ * @param arg3 (optional) A Node
+ * @param arg4 (optional) An offset number, or a PositionType
+ */
+export type Select = (
+    core: EditorCore,
+    arg1: Range | SelectionRangeEx | NodePosition | Node | SelectionPath | null,
+    arg2?: NodePosition | number | PositionType | TableSelection,
+    arg3?: Node,
+    arg4?: number | PositionType
+) => boolean;
+
+/**
  * Change the editor selection to the given range
  * @param core The EditorCore object
  * @param range The range to select
@@ -409,6 +428,17 @@ export interface CoreApiMap {
      * @param step Steps to move, can be 0, positive or negative
      */
     restoreUndoSnapshot: RestoreUndoSnapshot;
+
+    /**
+     * Select content according to the given information.
+     * There are a bunch of allowed combination of parameters. See IEditor.select for more details
+     * @param core The editor core object
+     * @param arg1 A DOM Range, or SelectionRangeEx, or NodePosition, or Node, or Selection Path
+     * @param arg2 (optional) A NodePosition, or an offset number, or a PositionType, or a TableSelection
+     * @param arg3 (optional) A Node
+     * @param arg4 (optional) An offset number, or a PositionType
+     */
+    select: Select;
 
     /**
      * Change the editor selection to the given range
