+++ packages/roosterjs-editor-types/lib/interface/index.ts
@@ -84,6 +84,7 @@ export {
     HasFocus,
     InsertNode,
     RestoreUndoSnapshot,
+    Select,
     SelectRange,
     SetContent,
     SwitchShadowEdit,
