+++ packages/roosterjs-editor-types/lib/type/CoreCreator.ts
@@ -1,0 +1,12 @@
+import EditorCore from '../interface/EditorCore';
+import EditorOptions from '../interface/EditorOptions';
+
+/**
+ * Type of Editor Core Creator
+ * @param contentDiv The DIV HTML element which will be the container element of editor
+ * @param options An optional options object to customize the editor
+ */
+export type CoreCreator<
+    TEditorCore extends EditorCore,
+    TEditorOptions extends EditorOptions = EditorOptions
+> = (contentDiv: HTMLDivElement, options: TEditorOptions) => TEditorCore;
